src/requirements.txt
pip<25.1
importlib-resources<6.5
absl-py==1.0.0+computecanada
cachetools==4.2.4+computecanada
certifi==2021.10.8+computecanada
charset-normalizer==2.0.12+computecanada
cycler==0.11.0+computecanada
fonttools==4.33.3+computecanada
google-auth==2.3.3+computecanada
google-auth-oauthlib==0.4.6+computecanada
grpcio==1.44.0+computecanada
idna==3.3+computecanada
imageio==2.16.2+computecanada
importlib-metadata==4.11.3+computecanada
joblib==1.1.0+computecanada
kiwisolver==1.3.2+computecanada
markdown==3.3.6+computecanada
matplotlib==3.5.1+computecanada
networkx==2.8.2+computecanada
nibabel==5.2.0+computecanada
numpy==1.22.2+computecanada
oauthlib==3.1.1+computecanada
opencv-python-headless==4.5.3.56+computecanada
packaging==21.3+computecanada
pillow==9.0.1+computecanada
Pillow-SIMD==7.0.0.post3+computecanada
protobuf==3.19.4+computecanada
py-spy==0.3.3+computecanada
pyasn1==0.4.8+computecanada
pyasn1-modules==0.2.8+computecanada
pyparsing==3.0.8+computecanada
python-dateutil==2.8.2+computecanada
pywavelets==1.3.0+computecanada
requests==2.27.1+computecanada
requests-oauthlib==1.3.0+computecanada
rsa==4.8+computecanada
scikit-image==0.18.2+computecanada
scikit-learn==1.0.2+computecanada
scipy==1.8.0+computecanada
six==1.16.0+computecanada
tensorboard==2.8.0+computecanada
tensorboard-data-server==0.6.1+computecanada
tensorboard-plugin-wit==1.8.0+computecanada
threadpoolctl==3.1.0+computecanada
tifffile==2022.4.8+computecanada
torch==1.10.0+computecanada
torchsummary==1.5.1+computecanada
torchvision==0.11.1+computecanada
typing-extensions==4.2.0+computecanada
urllib3==1.26.9+computecanada
werkzeug==2.1.2+computecanada
zipp==3.8.0+computecanada
